Refactor: move cas to a dedicated package
